add average ratio per day to bubble chart
in order to get a better idea of the average test code ratio, I am
adding a line into the bubble chart giving the average ratio. Because it
is unintuitive to indicate the average before the commits happen, I am
displaying the average for the day on the marker of the next day.

Also updated the Legend to update the line chart when any committers are
de/selected.