add dynamic timePeriod depending on commit time
I had the labels on the axis set to d3.time.days. This renders one day
per day between the last and first commit. This means, if your repo has
two commits, one from 2014-01-01 and one from 2013-01-01 you get 365
days rendered on the time axis This is barely readable.

I added some logic to determine the days between the first and last
commit and compare then to the available width on the chart. If that
would make for a crowded time axis, I am changing the timePeriod to
weeks or months or years, depending on the amount.