add chart Top Commits by Author and Size
this commit adds a new Bar Chart to Bubbly showing the top ten commits
per author sorted by size in percent. Percentage is calculated from the
top ten commits only and not from the sum of all commits. Percentage is
needed to prevent users with large commits to dwarf the bar charts of
the users with smaller sized commits.